+++ GRDB/Core/Row.swift
@@ -155,7 +155,7 @@ extension Row {
     /// fail, a fatal error is raised.
     public func value<Value: DatabaseValueConvertible>(atIndex index: Int) -> Value? {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
-        return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
+        return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
     }
     
     /// Returns the value at given index, converted to the requested type.
@@ -173,7 +173,7 @@ extension Row {
     public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(atIndex index: Int) -> Value? {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
         guard let sqliteStatement = sqliteStatement else {
-            return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
+            return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -187,7 +187,7 @@ extension Row {
     /// SQLite value can not be converted to `Value`.
     public func value<Value: DatabaseValueConvertible>(atIndex index: Int) -> Value {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
-        return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
+        return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
     }
     
     /// Returns the value at given index, converted to the requested type.
@@ -204,7 +204,7 @@ extension Row {
     public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(atIndex index: Int) -> Value {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
         guard let sqliteStatement = sqliteStatement else {
-            return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
+            return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -242,7 +242,7 @@ extension Row {
         guard let index = impl.index(ofColumn: columnName) else {
             return nil
         }
-        return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
+        return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
     }
     
     /// Returns the value at given column, converted to the requested type.
@@ -262,7 +262,7 @@ extension Row {
             return nil
         }
         guard let sqliteStatement = sqliteStatement else {
-            return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
+            return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -281,7 +281,7 @@ extension Row {
             // Programmer error
             fatalError("no such column: \(columnName)")
         }
-        return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
+        return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
     }
     
     /// Returns the value at given column, converted to the requested type.
@@ -303,7 +303,7 @@ extension Row {
             fatalError("no such column: \(columnName)")
         }
         guard let sqliteStatement = sqliteStatement else {
-            return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
+            return impl.databaseValue(atUncheckedIndex: index).losslessConvert()
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -874,7 +874,7 @@ private struct ArrayRowImpl : RowImpl {
     }
     
     func dataNoCopy(atUncheckedIndex index:Int) -> Data? {
-        return Data.convertOptional(from: databaseValue(atUncheckedIndex: index))
+        return databaseValue(atUncheckedIndex: index).losslessConvert()
     }
     
     func databaseValue(atUncheckedIndex index: Int) -> DatabaseValue {
@@ -926,7 +926,7 @@ private struct StatementCopyRowImpl : RowImpl {
     }
     
     func dataNoCopy(atUncheckedIndex index:Int) -> Data? {
-        return Data.convertOptional(from: databaseValue(atUncheckedIndex: index))
+        return databaseValue(atUncheckedIndex: index).losslessConvert()
     }
     
     func databaseValue(atUncheckedIndex index: Int) -> DatabaseValue {