+++ GRDB/Record/FetchedRecordsController.swift
@@ -436,8 +436,7 @@ private final class FetchedRecordsObserver<Record: RowConvertible> : Transaction
     }
     
     func observes(eventsOfKind eventKind: DatabaseEventKind) -> Bool {
-        // If impact is unknown, assume true
-        return eventKind.impacts(selectionInfo) ?? true
+        return eventKind.impacts(selectionInfo)
     }
     
     #if SQLITE_ENABLE_PREUPDATE_HOOK