+++ Tests/GRDBTests/DatabaseValueConvertibleFetchTests.swift
@@ -56,52 +56,6 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
         }
     }
 
-    func testFetchCursorConversionFailure() throws {
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.inDatabase { db in
-            func test(_ cursor: DatabaseCursor<Fetched>, sql: String) throws {
-                XCTAssertEqual(try cursor.next()!.int, 1)
-                do {
-                    _ = try cursor.next()
-                    XCTFail()
-                } catch let error as DatabaseError {
-                    XCTAssertEqual(error.resultCode, .SQLITE_ERROR)
-                    XCTAssertEqual(error.message, "could not convert database value NULL to \(Fetched.self)")
-                    XCTAssertEqual(error.sql!, sql)
-                    XCTAssertEqual(error.description, "SQLite error 1 with statement `\(sql)`: could not convert database value NULL to \(Fetched.self)")
-                }
-                do {
-                    _ = try cursor.next()
-                    XCTFail()
-                } catch let error as DatabaseError {
-                    XCTAssertEqual(error.resultCode, .SQLITE_ERROR)
-                    XCTAssertEqual(error.message, "could not convert database value \"foo\" to \(Fetched.self)")
-                    XCTAssertEqual(error.sql!, sql)
-                    XCTAssertEqual(error.description, "SQLite error 1 with statement `\(sql)`: could not convert database value \"foo\" to \(Fetched.self)")
-                }
-                XCTAssertEqual(try cursor.next()!.int, 2)
-                XCTAssertTrue(try cursor.next() == nil) // end
-            }
-            do {
-                let sql = "SELECT 1 UNION ALL SELECT NULL UNION ALL SELECT 'foo' UNION ALL SELECT 2"
-                let statement = try db.makeSelectStatement(sql)
-                try test(Fetched.fetchCursor(db, sql), sql: sql)
-                try test(Fetched.fetchCursor(statement), sql: sql)
-                try test(Fetched.fetchCursor(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).bound(to: Fetched.self).fetchCursor(db), sql: sql)
-            }
-            do {
-                let sql = "SELECT 0, 1 UNION ALL SELECT 0, NULL UNION ALL SELECT 0, 'foo' UNION ALL SELECT 0, 2"
-                let statement = try db.makeSelectStatement(sql)
-                let adapter = SuffixRowAdapter(fromIndex: 1)
-                try test(Fetched.fetchCursor(db, sql, adapter: adapter), sql: sql)
-                try test(Fetched.fetchCursor(statement, adapter: adapter), sql: sql)
-                try test(Fetched.fetchCursor(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).bound(to: Fetched.self).fetchCursor(db), sql: sql)
-            }
-        }
-    }
-
     func testFetchCursorStepFailure() throws {
         let dbQueue = try makeDatabaseQueue()
         let customError = NSError(domain: "Custom", code: 0xDEAD)
@@ -199,40 +153,6 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchAll(statement, adapter: adapter))
                 try test(Fetched.fetchAll(db, SQLRequest(sql, adapter: adapter)))
                 try test(SQLRequest(sql, adapter: adapter).bound(to: Fetched.self).fetchAll(db))
-            }
-        }
-    }
-
-    func testFetchAllConversionFailure() throws {
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.inDatabase { db in
-            func test(_ array: @autoclosure () throws -> [Fetched], sql: String) throws {
-                do {
-                    _ = try array()
-                    XCTFail()
-                } catch let error as DatabaseError {
-                    XCTAssertEqual(error.resultCode, .SQLITE_ERROR)
-                    XCTAssertEqual(error.message, "could not convert database value NULL to \(Fetched.self)")
-                    XCTAssertEqual(error.sql!, sql)
-                    XCTAssertEqual(error.description, "SQLite error 1 with statement `\(sql)`: could not convert database value NULL to \(Fetched.self)")
-                }
-            }
-            do {
-                let sql = "SELECT 1 UNION ALL SELECT NULL UNION ALL SELECT 2"
-                let statement = try db.makeSelectStatement(sql)
-                try test(Fetched.fetchAll(db, sql), sql: sql)
-                try test(Fetched.fetchAll(statement), sql: sql)
-                try test(Fetched.fetchAll(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).bound(to: Fetched.self).fetchAll(db), sql: sql)
-            }
-            do {
-                let sql = "SELECT 0, 1 UNION ALL SELECT 0, NULL UNION ALL SELECT 0, 2"
-                let statement = try db.makeSelectStatement(sql)
-                let adapter = SuffixRowAdapter(fromIndex: 1)
-                try test(Fetched.fetchAll(db, sql, adapter: adapter), sql: sql)
-                try test(Fetched.fetchAll(statement, adapter: adapter), sql: sql)
-                try test(Fetched.fetchAll(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).bound(to: Fetched.self).fetchAll(db), sql: sql)
             }
         }
     }
@@ -375,40 +295,6 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
         }
     }
 
-    func testFetchOneConversionFailure() throws {
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.inDatabase { db in
-            func test(_ value: @autoclosure () throws -> Fetched?, sql: String) throws {
-                do {
-                    _ = try value()
-                    XCTFail()
-                } catch let error as DatabaseError {
-                    XCTAssertEqual(error.resultCode, .SQLITE_ERROR)
-                    XCTAssertEqual(error.message, "could not convert database value \"foo\" to \(Fetched.self)")
-                    XCTAssertEqual(error.sql!, sql)
-                    XCTAssertEqual(error.description, "SQLite error 1 with statement `\(sql)`: could not convert database value \"foo\" to \(Fetched.self)")
-                }
-            }
-            do {
-                let sql = "SELECT 'foo'"
-                let statement = try db.makeSelectStatement(sql)
-                try test(Fetched.fetchOne(db, sql), sql: sql)
-                try test(Fetched.fetchOne(statement), sql: sql)
-                try test(Fetched.fetchOne(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).bound(to: Fetched.self).fetchOne(db), sql: sql)
-            }
-            do {
-                let sql = "SELECT 0, 'foo'"
-                let statement = try db.makeSelectStatement(sql)
-                let adapter = SuffixRowAdapter(fromIndex: 1)
-                try test(Fetched.fetchOne(db, sql, adapter: adapter), sql: sql)
-                try test(Fetched.fetchOne(statement, adapter: adapter), sql: sql)
-                try test(Fetched.fetchOne(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).bound(to: Fetched.self).fetchOne(db), sql: sql)
-            }
-        }
-    }
-
     func testFetchOneStepFailure() throws {
         let dbQueue = try makeDatabaseQueue()
         let customError = NSError(domain: "Custom", code: 0xDEAD)
@@ -501,44 +387,6 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchCursor(statement, adapter: adapter))
                 try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql, adapter: adapter)))
                 try test(SQLRequest(sql, adapter: adapter).bound(to: Optional<Fetched>.self).fetchCursor(db))
-            }
-        }
-    }
-
-    func testOptionalFetchCursorConversionFailure() throws {
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.inDatabase { db in
-            func test(_ cursor: DatabaseCursor<Fetched?>, sql: String) throws {
-                XCTAssertEqual(try cursor.next()!!.int, 1)
-                XCTAssertTrue(try cursor.next()! == nil)
-                do {
-                    _ = try cursor.next()
-                    XCTFail()
-                } catch let error as DatabaseError {
-                    XCTAssertEqual(error.resultCode, .SQLITE_ERROR)
-                    XCTAssertEqual(error.message, "could not convert database value \"foo\" to \(Fetched.self)")
-                    XCTAssertEqual(error.sql!, sql)
-                    XCTAssertEqual(error.description, "SQLite error 1 with statement `\(sql)`: could not convert database value \"foo\" to \(Fetched.self)")
-                }
-                XCTAssertEqual(try cursor.next()!!.int, 2)
-                XCTAssertTrue(try cursor.next() == nil) // end
-            }
-            do {
-                let sql = "SELECT 1 UNION ALL SELECT NULL UNION ALL SELECT 'foo' UNION ALL SELECT 2"
-                let statement = try db.makeSelectStatement(sql)
-                try test(Optional<Fetched>.fetchCursor(db, sql), sql: sql)
-                try test(Optional<Fetched>.fetchCursor(statement), sql: sql)
-                try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).bound(to: Optional<Fetched>.self).fetchCursor(db), sql: sql)
-            }
-            do {
-                let sql = "SELECT 0, 1 UNION ALL SELECT 0, NULL UNION ALL SELECT 0, 'foo' UNION ALL SELECT 0, 2"
-                let statement = try db.makeSelectStatement(sql)
-                let adapter = SuffixRowAdapter(fromIndex: 1)
-                try test(Optional<Fetched>.fetchCursor(db, sql, adapter: adapter), sql: sql)
-                try test(Optional<Fetched>.fetchCursor(statement, adapter: adapter), sql: sql)
-                try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).bound(to: Optional<Fetched>.self).fetchCursor(db), sql: sql)
             }
         }
     }
@@ -642,40 +490,6 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchAll(statement, adapter: adapter))
                 try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql, adapter: adapter)))
                 try test(SQLRequest(sql, adapter: adapter).bound(to: Optional<Fetched>.self).fetchAll(db))
-            }
-        }
-    }
-
-    func testOptionalFetchAllConversionFailure() throws {
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.inDatabase { db in
-            func test(_ array: @autoclosure () throws -> [Fetched?], sql: String) throws {
-                do {
-                    _ = try array()
-                    XCTFail()
-                } catch let error as DatabaseError {
-                    XCTAssertEqual(error.resultCode, .SQLITE_ERROR)
-                    XCTAssertEqual(error.message, "could not convert database value \"foo\" to \(Fetched.self)")
-                    XCTAssertEqual(error.sql!, sql)
-                    XCTAssertEqual(error.description, "SQLite error 1 with statement `\(sql)`: could not convert database value \"foo\" to \(Fetched.self)")
-                }
-            }
-            do {
-                let sql = "SELECT 1 UNION ALL SELECT NULL UNION ALL SELECT 'foo' UNION ALL SELECT 2"
-                let statement = try db.makeSelectStatement(sql)
-                try test(Optional<Fetched>.fetchAll(db, sql), sql: sql)
-                try test(Optional<Fetched>.fetchAll(statement), sql: sql)
-                try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).bound(to: Optional<Fetched>.self).fetchAll(db), sql: sql)
-            }
-            do {
-                let sql = "SELECT 0, 1 UNION ALL SELECT 0, NULL UNION ALL SELECT 0, 'foo' UNION ALL SELECT 0, 2"
-                let statement = try db.makeSelectStatement(sql)
-                let adapter = SuffixRowAdapter(fromIndex: 1)
-                try test(Optional<Fetched>.fetchAll(db, sql, adapter: adapter), sql: sql)
-                try test(Optional<Fetched>.fetchAll(statement, adapter: adapter), sql: sql)
-                try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).bound(to: Optional<Fetched>.self).fetchAll(db), sql: sql)
             }
         }
     }